added C code for direct matrix manipulation.
Imported matrix submodule so that I can explictly manage matrices,
and send the current projection/modelview matrix to the graphics
card.  This is in anticipation of upgrading to OpenGL 4.5.

I had previously assumed that glTranslate* and glRotate* sent
the parameters directly to the graphics card for accelerated
matrix multiplication, but since OpenGL 3.1+ require the programmer
to manipulate the projection/modelview matrix directly, I now
think my assumption was likely incorrect.